FIX: Move range chunk parallel code into threading
