+++ src/threading.rs
@@ -2,6 +2,8 @@
 /// Threading support functions and statics
 
 #[cfg(feature="threading")]
+use std::cmp::min;
+#[cfg(feature="threading")]
 use std::str::FromStr;
 #[cfg(feature="threading")]
 use once_cell::sync::Lazy;
@@ -12,7 +14,7 @@ pub use thread_tree::ThreadTree as ThreadPool;
 pub use thread_tree::ThreadTreeCtx as ThreadPoolCtx;
 
 use crate::kernel::GemmKernel;
-
+use crate::util::RangeChunk;
 
 /// Dummy threadpool
 #[cfg(not(feature="threading"))]
@@ -148,5 +150,118 @@ impl LoopThreadConfig {
     /// Number of packing buffers for A
     #[inline(always)]
     pub(crate) fn num_pack_a(&self) -> usize { self.loop3 as usize }
+}
+
+
+impl RangeChunk {
+    /// "Builder" method to create a RangeChunkParallel
+    pub(crate) fn parallel(self, nthreads: u8, pool: ThreadPoolCtx) -> RangeChunkParallel<fn()> {
+        fn nop() {}
+
+        RangeChunkParallel {
+            nthreads,
+            pool,
+            range: self,
+            thread_local: nop,
+        }
+    }
+}
+
+/// Intermediate struct for building the parallel execution of a range chunk.
+pub(crate) struct RangeChunkParallel<'a, G> {
+    range: RangeChunk,
+    nthreads: u8,
+    pool: ThreadPoolCtx<'a>,
+    thread_local: G,
+}
+
+impl<'a, G> RangeChunkParallel<'a, G> {
+    #[cfg(feature="threading")]
+    /// Set thread local setup function - called once per thread to setup thread local data.
+    pub(crate) fn thread_local<G2, R>(self, func: G2) -> RangeChunkParallel<'a, G2>
+        where G2: Fn(usize, usize) -> R + Sync
+    {
+        RangeChunkParallel {
+            nthreads: self.nthreads,
+            pool: self.pool,
+            thread_local: func,
+            range: self.range,
+        }
+    }
+
+    #[cfg(not(feature="threading"))]
+    /// Set thread local setup function - called once per thread to setup thread local data.
+    pub(crate) fn thread_local<G2, R>(self, func: G2) -> RangeChunkParallel<'a, G2>
+        where G2: FnOnce(usize, usize) -> R + Sync
+    {
+        RangeChunkParallel {
+            nthreads: self.nthreads,
+            pool: self.pool,
+            thread_local: func,
+            range: self.range,
+        }
+    }
+}
+
+#[cfg(not(feature="threading"))]
+impl<G, R> RangeChunkParallel<'_, G>
+    where G: FnOnce(usize, usize) -> R + Sync,
+{
+    pub(crate) fn for_each<F>(self, for_each: F)
+        where F: Fn(ThreadPoolCtx<'_>, &mut R, usize, usize) + Sync,
+    {
+        let mut local = (self.thread_local)(0, 1);
+        for (ln, chunk_size) in self.range {
+            for_each(self.pool, &mut local, ln, chunk_size)
+        }
+    }
+}
+
+
+#[cfg(feature="threading")]
+impl<G, R> RangeChunkParallel<'_, G>
+    where G: Fn(usize, usize) -> R + Sync,
+{
+    /// Execute loop iterations (parallel if enabled) using the given closure.
+    ///
+    /// The closure gets the following arguments for each iteration:
+    ///
+    /// - Thread pool context (used for child threads)
+    /// - Mutable reference to thread local data
+    /// - index of chunk (like RangeChunk)
+    /// - size of chunk (like RangeChunk)
+    pub(crate) fn for_each<F>(self, for_each: F)
+        where F: Fn(ThreadPoolCtx<'_>, &mut R, usize, usize) + Sync,
+    {
+        fn inner<F, G, R>(range: RangeChunk, index: usize, nthreads: usize, pool: ThreadPoolCtx<'_>,
+                          thread_local: G, for_each: F)
+            where G: Fn(usize, usize) -> R + Sync,
+                  F: Fn(ThreadPoolCtx<'_>, &mut R, usize, usize) + Sync
+        {
+            let mut local = thread_local(index, nthreads);
+            for (ln, chunk_size) in range.part(index, nthreads) {
+                for_each(pool, &mut local, ln, chunk_size)
+            }
+        }
+
+        debug_assert!(self.nthreads <= 4, "this method does not support nthreads > 4, got {}",
+                      self.nthreads);
+        let pool = self.pool;
+        let range = self.range;
+        let for_each = &for_each;
+        let local = &self.thread_local;
+        let nthreads = min(self.nthreads as usize, 4);
+        let f = move |ctx: ThreadPoolCtx<'_>, i| inner(range, i, nthreads, ctx, local, for_each);
+        if nthreads >= 4 {
+            pool.join4(&f);
+        } else if nthreads >= 3 {
+            pool.join3l(&f);
+        } else if nthreads >= 2 {
+            pool.join(|ctx| f(ctx, 0), |ctx| f(ctx, 1));
+        } else {
+            f(pool, 0)
+        }
+    }
+
 }
